Continue Desktop functionality

+++ shared/src/commonMain/kotlin/com/toasterofbread/spmp/ui/layout/radiobuilder/RadioBuilderPage.kt
@@ -108,7 +108,7 @@ fun RadioBuilderPage(
                         }
     
                         if (selected == null) {
-                            WaveBorder(Modifier.fillMaxWidth(), getOffset = { it })
+                            WaveBorder(Modifier.fillMaxWidth(), getOffset = { it.toInt() })
                         }
                     }
